renaming the globalConfigSlice
