+++ src/ml_board/frontend/Dashboard/src/components/topbar-with-drawer/TopBarWithDrawer.tsx
@@ -16,7 +16,7 @@ import * as React from 'react';
 import { useLocation, useNavigate } from "react-router-dom";
 import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import { RoutesMapping } from '../../app/RoutesMapping';
-import { changeTab, selectTab } from '../../redux/status/statusSlice';
+import { changeTab, selectTab } from '../../redux/globalConfig/globalConfigSlice';
 import { LogoOnly, LogoText } from "../../svgs_and_imgs/Icons";
 import Statistics from '../statistics/Statistics';
 // styles
